gooseberry_a721: Set display mode to hdmi
Since the gooseberry is a board stripped from a tablet, its fex file assumes
an lcd is present, but on the gooseberry non is connected, to lets activate
the hdmi output by default.

And for whom has heard me complain before that the hdmi on my gooseberry was
not working, this was a cable issue. I already suspected this as I had only
1 cable with a mini hdmi connector. I ordered a new one and now the issue is
gone.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>